- 重构 `env` 模块为 `storage`，提供更准确的存储管理功能命名 - 保持 `env` 向后兼容性，标记为弃用状态，建议迁移到 `storage` 模块
