+++ docs/CHANGELOG.md
@@ -10,6 +10,13 @@
 
 ---
 
+## [2.1.15-dev.2] - 2025/08/12
+### 变更
+- 重构 `env` 模块为 `storage`，提供更准确的存储管理功能命名
+- 保持 `env` 向后兼容性，标记为弃用状态，建议迁移到 `storage` 模块
+
+---
+
 ## [2.1.15-dev.1] - 2025/08/10
 ### 变更
 - 更新硬仓库地址
