+++ docs/api/ErisPulse/Core/logger.md
@@ -1,6 +1,6 @@
 # 📦 `ErisPulse.Core.logger` 模块
 
-<sup>自动生成于 2025-08-09 21:50:50</sup>
+<sup>自动生成于 2025-08-12 17:41:58</sup>
 
 ---
 
@@ -123,4 +123,4 @@ ErisPulse 日志系统
 
 ---
 
-<sub>文档最后更新于 2025-08-09 21:50:50</sub>
+<sub>文档最后更新于 2025-08-12 17:41:58</sub>
