+++ docs/api/ErisPulse/Core/mods.md
@@ -1,6 +1,6 @@
 # 📦 `ErisPulse.Core.mods` 模块
 
-<sup>自动生成于 2025-08-09 21:50:50</sup>
+<sup>自动生成于 2025-08-12 17:41:58</sup>
 
 ---
 
@@ -209,4 +209,4 @@ ErisPulse 模块管理器
 
 ---
 
-<sub>文档最后更新于 2025-08-09 21:50:50</sub>
+<sub>文档最后更新于 2025-08-12 17:41:58</sub>
