+++ docs/api/ErisPulse/Core/router.md
@@ -1,6 +1,6 @@
 # 📦 `ErisPulse.Core.router` 模块
 
-<sup>自动生成于 2025-08-09 21:50:50</sup>
+<sup>自动生成于 2025-08-12 17:41:58</sup>
 
 ---
 
@@ -108,4 +108,4 @@ ErisPulse 路由系统
 
 ---
 
-<sub>文档最后更新于 2025-08-09 21:50:50</sub>
+<sub>文档最后更新于 2025-08-12 17:41:58</sub>
