+++ docs/api/ErisPulse/__main__.md
@@ -1,6 +1,6 @@
 # 📦 `ErisPulse.__main__` 模块
 
-<sup>自动生成于 2025-08-09 21:50:50</sup>
+<sup>自动生成于 2025-08-12 17:41:58</sup>
 
 ---
 
@@ -314,4 +314,4 @@ ErisPulse命令行接口
 
 ---
 
-<sub>文档最后更新于 2025-08-09 21:50:50</sub>
+<sub>文档最后更新于 2025-08-12 17:41:58</sub>
